+++ src/components/application_manager/src/mobile_command_factory.cc
@@ -527,7 +527,7 @@ CommandSharedPtr MobileCommandFactory::CreateCommand(
     default: {
       command.reset(new commands::GenericResponse(message));
       (*message)[strings::params][strings::function_id] =
-          mobile_apis::FunctionID::GenericResponseID;
+          static_cast<int>(mobile_apis::FunctionID::GenericResponseID);
       break;
     }
   }
